# docs/swagger.yaml
basePath: /api
definitions:
  cmd_api_handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  github_com_yourusername_go-api-starter_cmd_api_models.DatabaseHealth:
    properties:
      error:
        type: string
      response_time:
        type: string
      status:
        type: string
    type: object
  github_com_yourusername_go-api-starter_cmd_api_models.HealthResponse:
    properties:
      database:
        $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.DatabaseHealth'
      service:
        type: string
      status:
        type: string
      timestamp:
        type: string
      uptime:
        type: string
      version:
        type: string
    type: object
  github_com_yourusername_go-api-starter_cmd_api_models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_yourusername_go-api-starter_cmd_api_models.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.User'
    type: object
  github_com_yourusername_go-api-starter_cmd_api_models.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  github_com_yourusername_go-api-starter_cmd_api_models.UpdateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
    required:
    - email
    - full_name
    type: object
  github_com_yourusername_go-api-starter_cmd_api_models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
  github_com_yourusername_go-api-starter_cmd_api_models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
  description: |-
    A production-ready Go REST API starter template with PostgreSQL, JWT Authentication,
    and best practices.
    **Features:**
    - ✅ Clean Architecture (Handler-Service-Repository)
    - ✅ JWT Authentication
    - ✅ Database Migrations
    - ✅ OpenAPI Documentation
  title: Go API Starter
  version: 1.0.0
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
      summary: Login user
      tags:
      - Authentication
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
  /api/health:
    get:
      consumes:
      - application/json
      description: Check if the API is running and database is accessible
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.HealthResponse'
        "503":
          description: Service unavailable if database is down
          schema:
            $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.HealthResponse'
      summary: Health check
      tags:
      - Health
  /api/users:
    get:
      consumes:
      - application/json
      description: Get a paginated list of users
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user account (admin only)
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - Users
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get a specific user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update a user's information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_yourusername_go-api-starter_cmd_api_models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"